requirements.txt
faiss-cpu==1.8.0
PyPDF2==3.0.1
python-dotenv==1.0.1
google-generativeai==0.7.2
tqdm==4.66.4
